Converted contact page to TS
